requirements.txt
pip==25.2
setuptools==75.8.0
wheel==0.45.1
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
triton==3.1.0
accelerate==1.10.0
transformers==4.49.0
peft==0.17.0
datasets==3.6.0
huggingface-hub==0.34.4
tokenizers==0.21.4
sentencepiece==0.2.0
bitsandbytes==0.46.1
xformers==0.0.28.post3
diffusers==0.34.0
safetensors==0.6.1
unsloth==2025.8.6
unsloth_zoo==2025.8.5
pandas==2.3.1
numpy==1.26.4
pyarrow==21.0.0
tqdm==4.67.1
pydantic==2.11.7
pydantic_core==2.33.2
protobuf==5.29.5
requests==2.32.4
urllib3==2.5.0
rich==14.1.0
tyro==0.9.28
termcolor==2.5.0
click==8.2.1
openai==1.99.9
google-generativeai==0.8.5
groq==0.31.0
python-dotenv==1.0.1
PyYAML==6.0.2
tenacity==8.5.0
matplotlib==3.10.5
langchain_community==0.3.13
cut-cross-entropy==25.1.1
flash-attn
vllm==0.7.3
deepspeed==0.17.4
liger-kernel==0.6.1
wandb==0.21.1
huggingface_hub[cli]
hf_transfer

# We are using a custom version of TRL for compatability with Zeng et al.'s Multi-Turn GRPO: pip install git+https://github.com/huggingface/trl.git@fc4dae256d924dfbb906af9c2e817bc6fb7b590b
git+https://github.com/huggingface/trl.git@fc4dae256d924dfbb906af9c2e817bc6fb7b590b
